backend/requirements.txt
fastapi
uvicorn
playwright
requests
google-generativeai
beautifulsoup4
lxml
python-dotenv
